fixed code coverage after converting from concerns to helpers
