--- app/helpers/app_security_helper.rb
@@ -1,16 +1,0 @@
-module AppSecurityHelper
-  def auth_check
-    if(logged_in_app? ==false)
-      if(controller_name != "app_security")
-        redirect_to app_firewall_path
-      end
-    end
-  end
-
-  def logged_in_app?
-    if(session["appauth"] == true)
-      return true
-    end
-    return false
-  end
-end
