+++ app/views/student_queues/index.html.erb
@@ -67,7 +67,6 @@
   <th>Name</th>
   <th>Course</th>
 </tr>
-
 <tr>
 <% @scheduled_queue.each do |entry| %>
       <tr id="q_row">
